refactor: remove legacy mouse helper and consolidate auto clicker functionality
