+++ autoClicker_cat/src/requirements.txt
@@ -1,0 +1,3 @@
+pydirectinput
+keyboard
+tkinter
